synthetic_proxy_eval/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synthetic-proxy-eval"
version = "0.1.0"
description = "Evaluation harness for proxy metrics on synthetic healthcare tabular data"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "pandas",
  "scipy",
  "scikit-learn",
  "joblib",
  "pyyaml",
  "matplotlib",
  "ucimlrepo",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["eval_harness"]
